tests/wpt/tests/css/css-mixins/mixin-conditionals.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS Mixins: Conditionals within mixins</title>
    <link rel="help" href="https://drafts.csswg.org/css-mixins-1/#defining-mixins">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>

    <style>
      @mixin --m1() {
        @result {
          color: red;
          --custom-property: wrong;
        }
        @media (width < 1000px) {
          @result {
            color: green;
          }
        }
        @supports (color: green) {
          @result {
            --custom-property: exists;
          }
        }
        @media (width > 1000px) {
          @result {
            color: red;
          }
        }
      }
      #e1 {
        @apply --m1;
      }
    </style>
    <div id="e1">This text should be green.</div>
    <script>
      test(() => {
        let e1 = document.getElementById('e1');
        assert_equals(getComputedStyle(e1).color, 'rgb(0, 128, 0)');
        assert_equals(getComputedStyle(e1).getPropertyValue('--custom-property'), 'exists');
      }, '@media and @supports wrapping @result');
    </script>


    <style>
      #container {
        container-type: size;
        width: 100px;
        height: 50px;
      }
      @mixin --m2() {
        @container (width = 50px) {
          @result {
            color: red;
          }
        }
        @container (width = 100px) {
          @result {
            color: green;
          }
        }
        @container (width = 200px) {
          @result {
            color: red;
          }
        }
      }
      #e2 {
        @apply --m2;
      }
    </style>
    <div id="container"><div id="e2">This text should be green.</div></div>
    <script>
      test(() => {
        let e2 = document.getElementById('e2');
        assert_equals(getComputedStyle(e2).color, 'rgb(0, 128, 0)');
      }, '@container wrapping @result');
    </script>


    <style>
      @mixin --m3() {
        --some-color: red;
        @media (width < 3000px) {
          --some-color: green;
        }
        @supports (color: green) {
          --some-other-color: blue;
        }
        @media (width > 3000px) {
          --some-color: red;
        }
        @result {
          color: var(--some-color);
          --custom-property: var(--some-other-color);
        }
      }
      #e3 {
        @apply --m3;
      }
    </style>
    <div id="e3">This text should be green.</div>
    <script>
      test(() => {
        let e3 = document.getElementById('e3');
        assert_equals(getComputedStyle(e3).color, 'rgb(0, 128, 0)');
        assert_equals(getComputedStyle(e3).getPropertyValue('--custom-property'), 'blue');
      }, '@media and @supports wrapping locals');
    </script>

  </body>
</html>
